Implemented Timecut Calculation logic.

+++ src/main/java/com/lukegjpotter/tools/bikeracetimecutcalculator/dto/RaceFinishRequestRecord.java
@@ -2,5 +2,5 @@ package com.lukegjpotter.tools.bikeracetimecutcalculator.dto;
 
 public record RaceFinishRequestRecord(
         String raceDuration,
-        Double percentageTimeCut) {
+        int percentageTimeCut) {
 }
